+++ tests/testthat/test-bake.R
@@ -29,34 +29,34 @@ test_that("bake() and juice() doens't turn strings into factors #317", {
   exp_data <- tibble(f1 = factor(1), f2 = "1", c1 = "1")
 
   expect_identical(
-    recipe(~., exp_data) %>%
-      prep(strings_as_factors = TRUE) %>%
+    recipe(~., exp_data, strings_as_factors = TRUE) %>%
+      prep() %>%
       juice(),
     tibble(f1 = factor(1), f2 = factor(1), c1 = factor(1))
   )
 
   # juice()
   expect_identical(
-    recipe(~., exp_data) %>%
+    recipe(~., exp_data, strings_as_factors = TRUE) %>%
       step_factor2string(f1) %>%
-      prep(strings_as_factors = TRUE) %>%
+      prep() %>%
       juice(),
     tibble(f1 = "1", f2 = factor(1), c1 = factor(1))
   )
 
   # bake(new_data = NULL)
   expect_identical(
-    recipe(~., exp_data) %>%
-      prep(strings_as_factors = TRUE) %>%
+    recipe(~., exp_data, strings_as_factors = TRUE) %>%
+      prep() %>%
       bake(new_data = NULL),
     tibble(f1 = factor(1), f2 = factor(1), c1 = factor(1))
   )
 
   # bake(new_data = data)
   expect_identical(
-    recipe(~., exp_data) %>%
+    recipe(~., exp_data, strings_as_factors = TRUE) %>%
       step_factor2string(f1) %>%
-      prep(strings_as_factors = TRUE) %>%
+      prep() %>%
       bake(new_data = exp_data),
     tibble(f1 = "1", f2 = factor(1), c1 = factor(1))
   )