+++ tests/testthat/test-basics.R
@@ -40,19 +40,20 @@ test_that("Recipe on missspelled variables in formulas", {
 })
 
 test_that("return character or factor values", {
-  raw_recipe <- recipe(HHV ~ ., data = biomass)
-  centered <- raw_recipe %>%
+  rec_spec <- recipe(HHV ~ ., data = biomass, strings_as_factors = FALSE) %>%
     step_center(carbon, hydrogen, oxygen, nitrogen, sulfur)
 
   centered_char <- prep(
-    centered,
-    training = biomass,
-    strings_as_factors = FALSE
+    rec_spec,
+    training = biomass
   )
   char_var <- bake(centered_char, new_data = head(biomass))
   expect_equal(class(char_var$sample), "character")
 
-  centered_fac <- prep(centered, training = biomass, strings_as_factors = TRUE)
+  rec_spec <- recipe(HHV ~ ., data = biomass, strings_as_factors = TRUE) %>%
+    step_center(carbon, hydrogen, oxygen, nitrogen, sulfur)
+
+  centered_fac <- prep(rec_spec, training = biomass)
   fac_var <- bake(centered_fac, new_data = head(biomass))
   expect_equal(class(fac_var$sample), "factor")
   expect_equal(levels(fac_var$sample), sort(unique(biomass$sample)))
@@ -530,5 +531,93 @@ test_that("recipe() error for table input (#1416)", {
   expect_snapshot(
     error = TRUE,
     recipe(Titanic, Survived ~ .)
+  )
+})
+
+
+test_that("precedence for strings_as_factors in `recipe()`", {
+  local_options(lifecycle_verbosity = "quiet")
+
+  # Takes precedence over value in `prep()`
+  string_recipe <- recipe(HHV ~ ., data = biomass, strings_as_factors = FALSE)
+  prepped_string_recipe <- prep(
+    string_recipe,
+    training = biomass,
+    strings_as_factors = TRUE
+  )
+
+  factor_recipe <- recipe(HHV ~ ., data = biomass, strings_as_factors = TRUE)
+  prepped_factor_recipe <- prep(
+    factor_recipe,
+    training = biomass,
+    strings_as_factors = FALSE
+  )
+
+  char_var <- bake(prepped_string_recipe, new_data = head(biomass))
+  expect_identical(class(char_var$sample), "character")
+
+  factor_var <- bake(prepped_factor_recipe, new_data = head(biomass))
+  expect_identical(class(factor_var$sample), "factor")
+
+  # `prep()` takes precedence if it isn't set in `recipe()`
+  string_recipe <- recipe(HHV ~ ., data = biomass)
+
+  prepped_string_recipe <- prep(
+    string_recipe,
+    training = biomass,
+    strings_as_factors = FALSE
+  )
+
+  char_var <- bake(prepped_string_recipe, new_data = head(biomass))
+  expect_identical(class(char_var$sample), "character")
+})
+
+test_that("strings_as_factors in `recipe()` for different roles", {
+  # outcomes
+  # predictors
+  # id / other
+  # undeclared
+
+  ex_dat <- tibble(
+    outcome_string = letters,
+    outcome_factor = factor(letters),
+    predictor_string = letters,
+    predictor_factor = factor(letters),
+    id_string = letters,
+    id_factor = factor(letters),
+    undeclared_string = letters,
+    undeclared_factor = factor(letters)
+  )
+
+  res_false <- recipe(ex_dat, strings_as_factors = FALSE) %>%
+    update_role(starts_with("outcome"), new_role = "outcome") %>%
+    update_role(starts_with("predictor"), new_role = "predictor") %>%
+    update_role(starts_with("id"), new_role = "id") %>%
+    prep() %>%
+    bake(NULL)
+
+  expect_identical(vctrs::vec_ptype(ex_dat), vctrs::vec_ptype(res_false))
+
+  res_true <- recipe(ex_dat, strings_as_factors = TRUE) %>%
+    update_role(starts_with("outcome"), new_role = "outcome") %>%
+    update_role(starts_with("predictor"), new_role = "predictor") %>%
+    update_role(starts_with("id"), new_role = "id") %>%
+    prep() %>%
+    bake(NULL)
+
+  exp_true_classes <- list(
+    outcome_string = "factor",
+    outcome_factor = "factor",
+    predictor_string = "factor",
+    predictor_factor = "factor",
+    id_string = "character",
+    id_factor = "factor",
+    undeclared_string = "character",
+    undeclared_factor = "factor"
+  )
+
+  expect_identical(
+    lapply(res_true, class),
+    exp_true_classes
   )
 })