+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * Make it so `recipe.formula()` can't take table objects as input, in accordance with documentation. (#1416)
 
+* The `strings_as_factors` argument of `prep.recipe()` has been soft-deprecated in favor of `recipe(strings_as_factors)`. If both are provided, the value in `recipe()` takes precedence. This allows control of recipe behavior within a workflow, which wasn't previously possible. (@smingerson, #331, #287)
+
 * Fixed bug where `step_lincomb()` would remove both variables if they were identical. (#1357)
 
 * `step_spline_b()`, `step_spline_convex()`, `step_spline_monotone()`, `step_spline_natural()`, and `step_spline_nonnegative()` now gives informative errors when applied to zero variance predictors. (#1455)